+++ src/main/java/kitchenpos/enums/OrderStatus.java
@@ -1,4 +1,4 @@
-package kitchenpos.domain;
+package kitchenpos.enums;
 
 public enum OrderStatus {
     COOKING, MEAL, COMPLETION
