+++ src/main/java/kitchenpos/order/application/OrderService.java
@@ -1,9 +1,8 @@
 package kitchenpos.order.application;
 
 import kitchenpos.menu.application.MenuService;
-import kitchenpos.order.domain.Orders;
-import kitchenpos.order.domain.OrderLineItemRepository;
 import kitchenpos.order.domain.OrderRepository;
+import kitchenpos.order.domain.Orders;
 import kitchenpos.order.dto.OrderRequest;
 import kitchenpos.order.dto.OrderResponse;
 import kitchenpos.table.application.TableService;
