+++ src/main/java/kitchenpos/order/dto/OrderRequest.java
@@ -1,6 +1,6 @@
 package kitchenpos.order.dto;
 
-import kitchenpos.domain.OrderStatus;
+import kitchenpos.enums.OrderStatus;
 import kitchenpos.order.domain.Orders;
 import kitchenpos.order.domain.OrderLineItem;
